\docType{package}
\name{bunyan}
\alias{bunyan}
\alias{bunyan-package}
\title{bunyan}
\description{
  Joyent bunyan JSON error logging package
}
\details{
  UNDER DEVELOPMENT

  R functions to log errors in structured JSON

  ...
}
\references{
  http://blog.nodejs.org/2012/03/28/service-logging-in-json-with-bunyan/
}

