\name{bunyanBuffer}
\alias{bunyanBuffer}
\title{Shows contents of bunyan memory circular buffer
First/Last entry can be determined from timestamp}
\usage{
  bunyanBuffer()
}
\description{
  Shows contents of bunyan memory circular buffer
  First/Last entry can be determined from timestamp
}
\keyword{bunyan,}

