\name{bunyanTracebackN}
\alias{bunyanTracebackN}
\title{Removes Manta object specified by full manta path or from current
working manta directory}
\usage{
  bunyanTracebackN()
}
\description{
  Removes Manta object specified by full manta path or from
  current working manta directory
}
\keyword{bunyan,}
\keyword{setpoint}

