\name{bunyanClearSetpoint}
\alias{bunyanClearSetpoint}
\title{Clears the memory buffer setpoint after which
if no entries above level threshold, bunyanTraceback
and bunyanTracebackN will return blank and 0 results}
\usage{
  bunyanClearSetpoint()
}
\description{
  Clears the memory buffer setpoint after which if no
  entries above level threshold, bunyanTraceback and
  bunyanTracebackN will return blank and 0 results
}
\keyword{bunyan,}
\keyword{setpoint}

