\name{bunyanSetpoint}
\alias{bunyanSetpoint}
\title{Removes Manta object specified by full manta path or from current
working manta directory}
\usage{
  bunyanSetpoint()
}
\description{
  Removes Manta object specified by full manta path or from
  current working manta directory
}
\keyword{bunyan,}
\keyword{setpoint}

